solver/kagamiz/pyproject.toml
[project]
name = "kagamiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
